BUG: DICOMRadionuclideHalfLife should be treated as double

+++ starviewer/src/core/decaycorrectionfactorformulacalculator.h
@@ -45,7 +45,7 @@ private:
     
     /// (0018,1075) Type 3. The radionuclide half life, in seconds, that was used in the correction of this image.
     /// Contained in Radiopharmaceutical Information Sequence (0054,0016)
-    int m_radionuclideHalfLifeInSeconds;
+    double m_radionuclideHalfLifeInSeconds;
     
     /// Time lapse used in the formula.
     /// If Decay Correction = START, timeLapse = Series Time - Radiopharmaceutical Start Time (0018,1072)
